+++ Demo/MainWindow.xaml.cs
@@ -55,6 +55,11 @@ public partial class MainWindow : INotifyPropertyChanged
             new ChapterWrapPanelControl()
         ];
 
+        _statusAndInfo =
+        [
+            new ChapterBadgeControl()
+        ];
+
         Controls = _basicInput;
 
         DataContext = this;
@@ -64,6 +69,7 @@ public partial class MainWindow : INotifyPropertyChanged
     private List<UserControl> _collections { get; }
     private List<UserControl> _text { get; }
     private List<UserControl> _layout { get; }
+    private List<UserControl> _statusAndInfo { get; }
 
     public List<UserControl> Controls { get; private set; }
 
@@ -87,6 +93,9 @@ public partial class MainWindow : INotifyPropertyChanged
                 break;
             case "Layout":
                 Controls = _layout;
+                break;
+            case "StatusAndInfo":
+                Controls = _statusAndInfo;
                 break;
         }